+++ sample/frontend/src/containers/IframePage/index.tsx
@@ -10,7 +10,7 @@ import { v4 as uuidv4 } from 'uuid';
 import { createAccessTokenSetting, fetchAccessToken } from '@/api';
 import ErrorDialog from '@/components/ErrorDialog';
 import { DEFAULT_LAYOUT, LS_CLIENT_ID, LS_CONF_URL, ROOM_CONFIG, SIGNALING_URL, SUBVIEW_CONFIG, THEME_CONFIG, THETA_ZOOM_MAX_RANGE, TOOLBAR_CONFIG } from '@/constants';
-import LSConferenceIframe, { ConnectOptions, CreateParameters } from '@/lib/ls-conf-sdk';
+import LSConferenceIframe, { ConnectOptions, CreateParameters, LSConfError, LSConfErrorEvent } from '@/lib/ls-conf-sdk';
 
 const CREATE_PARAMETERS: CreateParameters = {
   defaultLayout: (DEFAULT_LAYOUT as 'gallery' | 'presentation' | 'fullscreen') || undefined,
@@ -45,21 +45,29 @@ const IframePage: React.FC<Record<string, never>> = () => {
   const iframeContainerRef = useRef<HTMLDivElement>(null);
   const [lsConfIframe, setLsConfIframe] = useState<LSConferenceIframe | null>(null);
   const [errorMessage, setErrorMessage] = useState<string | null>(null);
+  // ダイアログで表示中のダミーデバイス変更通知のデバイス種別
+  const showingDummyDeviceTypes = useRef<string[]>([]);
   const showErrorDialog = errorMessage !== null;
   const connectionId: string = uuidv4();
   const audioMuteType = audio_mute_type && (audio_mute_type === 'soft' || audio_mute_type === 'hard') ? audio_mute_type : undefined;
+  let isUseDummyDevice = Boolean(use_dummy_device && typeof use_dummy_device === 'string' && use_dummy_device.toLowerCase() === 'true');
   const videoCodec = video_codec && (video_codec === 'h264' || video_codec === 'vp8' || video_codec === 'vp9' || video_codec === 'h265' || video_codec === 'av1') ? video_codec : undefined;
   const iceServersProtocol =
     ice_servers_protocol && (ice_servers_protocol === 'all' || ice_servers_protocol === 'udp' || ice_servers_protocol === 'tcp' || ice_servers_protocol === 'tls') ? ice_servers_protocol : undefined;
-  const downloadLog = async (iframe: LSConferenceIframe, errorEvent?: ErrorEvent): Promise<void> => {
+  const downloadLog = async (iframe: LSConferenceIframe, e?: LSConfError | LSConfErrorEvent): Promise<void> => {
     let log = 'LSConfSample Log\n\n';
-    if (errorEvent) {
+    if (e instanceof LSConfError) {
       log += `********** Error Message **********\n`;
-      log += `${errorEvent.message}\n`;
+      log += `${e.message}\n`;
+      log += `********** toReportString() *******\n`;
+      log += `${e.toReportString()}\n`;
+    } else if (e instanceof LSConfErrorEvent) {
+      log += `********** Error Message **********\n`;
+      log += `${e.message}\n`;
+      log += `********** toReportString() *******\n`;
+      log += `${e.error.toReportString()}\n`;
     }
     log += `********** ApplicationLog *********\n`;
-    // eslint-disable-next-line @typescript-eslint/no-var-requires
-    log += `LSConfSample Version: v${require('../../../../frontend/package.json').version}\n`;
     log += `LSConfURL: ${LS_CONF_URL || 'default'}\n`;
     log += `LSClientID: ${LS_CLIENT_ID || 'unknown'}\n`;
     log += `SignalingURL: ${SIGNALING_URL || 'default'}\n`;
@@ -69,10 +77,6 @@ const IframePage: React.FC<Record<string, never>> = () => {
       log += `${await iframe.getLSConfLog()}\n`;
     } catch {
       log += `Failed to getLSConfLog.\n`;
-    }
-    if (errorEvent && errorEvent.error && errorEvent.error.toReportString) {
-      log += `********** toReportString *********\n`;
-      log += errorEvent.error.toReportString;
     }
     const downLoadLink = document.createElement('a');
     downLoadLink.download = `ls-conf-sample_${format(new Date(), 'yyyyMMdd_HHmmss')}.log`;
@@ -105,15 +109,11 @@ const IframePage: React.FC<Record<string, never>> = () => {
   };
   const createAndConnectRoom = async (): Promise<void> => {
     if (!username || !roomId || typeof username !== 'string') {
-      // 現在 ls-conf-sdk への対応と同様にエラーをそのまま errorMessage に入れている
-      // TODO(kdxu): ls-conf-sdk のシステムエラーに対するユーザーへのメッセージが仕様として策定され次第、こちらのエラー文言も合わせて修正する
-      setErrorMessage('INVALID-PARAMETERS');
+      setErrorMessage('入室パラメータが不正です');
       return;
     }
     if (!iframeContainerRef.current) {
-      // 現在 ls-conf-sdk への対応と同様にエラーをそのまま errorMessage に入れている
-      // TODO(kdxu): ls-conf-sdk のシステムエラーに対するユーザーへのメッセージが仕様として策定され次第、こちらのエラー文言も合わせて修正する
-      setErrorMessage('INVALID-IFRAME-CONTAINER');
+      setErrorMessage('iframeContainerが不正です');
       return;
     }
     let iframe: LSConferenceIframe;
@@ -125,7 +125,11 @@ const IframePage: React.FC<Record<string, never>> = () => {
       }
       iframe = await LSConferenceIframe.create(iframeContainerRef.current, CREATE_PARAMETERS);
     } catch (e) {
-      setErrorMessage(e.message);
+      if (e instanceof LSConfError) {
+        setErrorMessage(e.message);
+      } else {
+        setErrorMessage(`Unexpected error occurred: ${e}`);
+      }
       return;
     }
     iframe.onShareRequested(async () => {
@@ -135,7 +139,11 @@ const IframePage: React.FC<Record<string, never>> = () => {
         const accessTokenSetting = createAccessTokenSetting(roomId, screenShareConnectionId, bitrate_reservation_mbps, room_type, max_connections);
         screenShareAccessToken = await fetchAccessToken(accessTokenSetting);
       } catch (e) {
-        setErrorMessage(e.message);
+        if (e instanceof Error) {
+          setErrorMessage(`アクセストークンの取得に失敗しました: ${e.message}`);
+        } else {
+          setErrorMessage(`Unexpected error occurred: ${e}`);
+        }
         return;
       }
       return screenShareAccessToken;
@@ -148,12 +156,12 @@ const IframePage: React.FC<Record<string, never>> = () => {
       maxVideoBitrate: Number(video_bitrate),
       maxShareBitrate: Number(share_bitrate),
       // eslint-disable-next-line @typescript-eslint/naming-convention
-      useDummyDevice: Boolean(use_dummy_device && typeof use_dummy_device === 'string' && use_dummy_device.toLowerCase() === 'true'),
+      useDummyDevice: isUseDummyDevice,
       signalingURL: SIGNALING_URL,
       videoCodec: videoCodec,
       iceServersProtocol,
     };
-    iframe.addEventListener('error', async (e: ErrorEvent) => {
+    iframe.addEventListener('error', async (e: LSConfErrorEvent) => {
       // TODO(hase): ChromeのMediaRecorderのバグの暫定対応
       // 既知の問題のNo.23の現象を検知した時にエラー表示を行う。回避方法などの詳細は以下のリンクをご覧ください。
       // cf: https://api.livestreaming.ricoh/document/ricoh-live-streaming-conference-%e6%97%a2%e7%9f%a5%e3%81%ae%e5%95%8f%e9%a1%8c/
@@ -182,7 +190,11 @@ const IframePage: React.FC<Record<string, never>> = () => {
       try {
         await iframe.addRecordingMember(targetSubview, connectionId);
       } catch (e) {
-        console.warn(`Failed to addRecordingMember in startRecording event. Detail: ${JSON.stringify(e.detail)}`);
+        if (e instanceof LSConfError) {
+          console.warn(`Failed to addRecordingMember in startRecording event. Detail: ${JSON.stringify(e.detail)}`);
+        } else {
+          console.warn(`Unexpected error occurred: ${e}`);
+        }
       }
     });
     iframe.addEventListener('stopRecording', async (e: CustomEvent) => {
@@ -191,8 +203,24 @@ const IframePage: React.FC<Record<string, never>> = () => {
       try {
         await iframe.removeRecordingMember(targetSubview, connectionId);
       } catch (e) {
-        console.warn(`Failed to removeRecordingMember in stopRecording event. Detail: ${JSON.stringify(e.detail)}`);
+        if (e instanceof LSConfError) {
+          console.warn(`Failed to removeRecordingMember in stopRecording event. Detail: ${JSON.stringify(e.detail)}`);
+        } else {
+          console.warn(`Unexpected error occurred: ${e}`);
+        }
       }
+    });
+    iframe.addEventListener('mediaDeviceChanged', async (e: CustomEvent) => {
+      // ダミーデバイスに切り替わった場合はダイアログを表示する
+      if (e.detail.deviceId === 'dummy-device') {
+        const deviceTypeName = e.detail.kind === 'videoinput' ? 'カメラ' : e.detail.kind === 'audioinput' ? 'マイク' : 'スピーカー';
+        if (!showingDummyDeviceTypes.current.some((deviceType) => deviceType === deviceTypeName)) {
+          showingDummyDeviceTypes.current.push(deviceTypeName);
+        }
+        const deviceTypeNames = showingDummyDeviceTypes.current.join('と');
+        setErrorMessage(`${deviceTypeNames}が接続されていません。${deviceTypeNames}を接続し直し、デバイス設定ダイアログから再度${deviceTypeNames}を設定してください。`);
+      }
+      console.log(`mediaDeviceChanged: ${JSON.stringify(e.detail)}`);
     });
     // ツールバーのダウンロードボタンを押した場合はログをダウンロードする
     iframe.addApplicationEventListener('log', async () => {
@@ -204,41 +232,56 @@ const IframePage: React.FC<Record<string, never>> = () => {
       }
     });
     try {
-      // join APIの実行中にデバイスアクセス許可ダイアログが表示されてJoinTimeoutが発生することを防ぐため、
-      // 事前にデバイスのアクセス許可を取得してからAccessTokenの生成とjoin APIの実行を行う
-      // Safariは60秒経過でデバイスのアクセス許可が取り消されるため、この処理から60秒以内にjoin APIを実行すること
-      await iframe.getMediaDevices();
-    } catch (e) {
-      setErrorMessage(e.message);
-      try {
-        await downloadLog(iframe, e);
-      } catch {
-        console.warn('Failed to download log.');
+      if (!isUseDummyDevice) {
+        // join APIの実行中にデバイスアクセス許可ダイアログが表示されてJoinTimeoutが発生することを防ぐため、
+        // 事前にデバイスのアクセス許可を取得してからAccessTokenの生成とjoin APIの実行を行う
+        // Safariは60秒経過でデバイスのアクセス許可が取り消されるため、この処理から60秒以内にjoin APIを実行すること
+        await iframe.getMediaDevices();
       }
-      return;
+    } catch (e) {
+      // ダミーデバイスでの入室に切り替える
+      isUseDummyDevice = true;
+      if (e instanceof LSConfError) {
+        console.warn(`Failed to getMediaDevices. Join with dummy device. error: ${JSON.stringify(e.detail)}`);
+      } else {
+        console.warn(`Unexpected error occurred: ${e}`);
+      }
     }
     let accessToken;
     try {
       const accessTokenSetting = createAccessTokenSetting(roomId, connectionId, bitrate_reservation_mbps, room_type, max_connections);
       accessToken = await fetchAccessToken(accessTokenSetting);
     } catch (e) {
-      setErrorMessage(e.message);
+      if (e instanceof Error) {
+        setErrorMessage(`アクセストークンの取得に失敗しました: ${e.message}`);
+      } else {
+        setErrorMessage(`Unexpected error occurred: ${e}`);
+      }
       return;
     }
     try {
       await iframe.join(LS_CLIENT_ID, accessToken, connectOptions);
     } catch (e) {
-      setErrorMessage(e.message);
-      try {
-        await downloadLog(iframe, e);
-      } catch {
-        console.warn('Failed to download log.');
+      if (e instanceof LSConfError) {
+        if (e.detail.code === 4120) {
+          setErrorMessage('デバイスの取得に失敗したため、カメラとマイクを使用せずに参加します。マイクデバイスへのアクセス許可がない場合、スピーカーから音が出ない場合があります。');
+        } else {
+          setErrorMessage(e.message);
+          try {
+            await downloadLog(iframe, e);
+          } catch {
+            console.warn('Failed to download log.');
+          }
+        }
+      } else {
+        setErrorMessage(`Unexpected error occurred: ${e}`);
       }
       return;
     }
     setLsConfIframe(iframe);
   };
   const onCloseErrorDialog = (): void => {
+    showingDummyDeviceTypes.current = [];
     setErrorMessage(null);
   };
   useLayoutEffect(() => {