+++ sample/frontend/webpack.config.js
@@ -268,14 +268,6 @@ const configSchema = {
                 },
               }
             },
-            "dialog": {
-              "type": "object",
-              "properties": {
-                "inputFocusColor": {
-                  "type": "string",
-                },
-              }
-            },
           }
         }
       }